Allow usage from python and pure rust
