+++ pyproject.toml
@@ -27,7 +27,7 @@ dependencies = [
 packages = ["twisterl"]
 
 [tool.maturin]
-features = ["pyo3/extension-module"]
+features = ["pyo3/extension-module", "python_bindings"]
 profile = "release"
 module-name = "twisterl.twisterl_rs"
