+++ rust/src/nn/policy.rs
@@ -11,7 +11,6 @@ copyright notice, and modified files need to carry a notice indicating
 that they have been altered from the originals.
 */
 
-use pyo3::prelude::*;
 use rand::{prelude::Distribution, Rng};
 
 use crate::nn::modules::Sequential;
