+++ rust/src/python_interface/mod.rs
@@ -18,3 +18,4 @@ pub mod layers;
 pub mod modules;
 pub mod policy;
 pub mod pyenv;
+pub mod python_bindings;
